[#1481] Move for_each_field to AtomicCopy trait
