+++ iceoryx2-bb/derive-macros/tests-common/src/zero_copy_send_tests.rs
@@ -12,7 +12,6 @@
 
 extern crate alloc;
 
-use alloc::vec::Vec;
 use iceoryx2_bb_derive_macros::ZeroCopySend;
 use iceoryx2_bb_elementary_traits::zero_copy_send::ZeroCopySend;
 use iceoryx2_bb_testing::assert_that;
@@ -188,222 +187,4 @@ pub fn for_unions() {
     let sut = BasicUnionTest { _val1: 12 };
     assert_that!(is_zero_copy_send(&sut), eq true);
     assert_that!(unsafe { BasicUnionTest::type_name() }, eq "TryMadHoney");
-}
-
-#[test]
-#[should_panic]
-pub fn field_offsets_and_sizes_cannot_be_calculated_for_unions() {
-    let sut = BasicUnionTest { _val1: 12 };
-    sut.__for_each_field(&mut |_, _| {});
-}
-
-#[test]
-#[should_panic]
-pub fn field_offsets_and_sizes_cannot_be_calculated_for_generic_unions() {
-    #[repr(C)]
-    #[derive(ZeroCopySend)]
-    union GenericUnion<T: Copy + ZeroCopySend> {
-        val1: u8,
-        val2: T,
-    }
-
-    let sut = GenericUnion { val2: 0u64 };
-    sut.__for_each_field(&mut |_, _| {});
-}
-
-#[test]
-#[should_panic]
-pub fn field_offsets_and_sizes_cannot_be_calculated_for_unions_with_struct_field() {
-    #[repr(C)]
-    #[derive(Clone, Copy, ZeroCopySend)]
-    struct SomeNamedStruct {
-        a: u8,
-        b: u64,
-    }
-
-    #[repr(C)]
-    #[derive(ZeroCopySend)]
-    union NestedUnion {
-        val1: u8,
-        val2: SomeNamedStruct,
-    }
-
-    let sut = NestedUnion {
-        val2: SomeNamedStruct { a: 0, b: 0 },
-    };
-    sut.__for_each_field(&mut |_, _| {});
-}
-
-#[test]
-#[should_panic]
-pub fn field_offsets_and_sizes_cannot_be_calculated_for_structs_with_union() {
-    #[repr(C)]
-    #[derive(ZeroCopySend)]
-    struct NestedStruct {
-        a: BasicUnionTest,
-        b: u8,
-    }
-
-    let sut = NestedStruct {
-        a: BasicUnionTest { _val1: 12 },
-        b: 0,
-    };
-    sut.__for_each_field(&mut |_, _| {});
-}
-
-#[test]
-#[should_panic]
-pub fn field_offsets_and_sizes_cannot_be_calculated_for_structs_with_enum() {
-    #[repr(C)]
-    #[derive(ZeroCopySend)]
-    enum SomeEnum {
-        A,
-        B,
-    }
-
-    #[repr(C)]
-    #[derive(ZeroCopySend)]
-    struct NestedStruct {
-        a: SomeEnum,
-        b: u8,
-    }
-
-    let sut = NestedStruct {
-        a: SomeEnum::A,
-        b: 0,
-    };
-    sut.__for_each_field(&mut |_, _| {});
-}
-
-#[test]
-pub fn field_offsets_and_sizes_are_correct_for_named_struct() {
-    let mut v = Vec::new();
-    let sut = NamedTestStruct {
-        _val1: 0,
-        _val2: Foo(0),
-    };
-    sut.__for_each_field(&mut |offset, size| {
-        v.push((offset, size));
-    });
-
-    assert_that!(v, len 2);
-    assert_that!(v[0], eq(0, 8));
-    assert_that!(v[1], eq(8, 2));
-}
-
-#[test]
-pub fn field_offsets_and_sizes_are_correct_for_generic_named_struct() {
-    let mut v = Vec::new();
-    let sut = GenericNamedTestStruct {
-        _val1: 0u8,
-        _val2: 0i32,
-    };
-    sut.__for_each_field(&mut |offset, size| {
-        v.push((offset, size));
-    });
-
-    assert_that!(v, len 2);
-    assert_that!(v[0], eq(0, 1));
-    assert_that!(v[1], eq(4, 4));
-}
-
-#[test]
-pub fn field_offsets_and_sizes_are_correct_for_unnamed_struct() {
-    let mut v = Vec::new();
-    let sut = UnnamedTestStruct(0, 0, Foo(0));
-    sut.__for_each_field(&mut |offset, size| {
-        v.push((offset, size));
-    });
-
-    assert_that!(v, len 3);
-    assert_that!(v[0], eq(0, 4));
-    assert_that!(v[1], eq(8, 8));
-    assert_that!(v[2], eq(16, 2));
-}
-
-#[test]
-pub fn field_offsets_and_sizes_are_correct_for_generic_unnamed_struct() {
-    let mut v = Vec::new();
-    let sut = GenericUnnamedTestStruct(0u64, UnnamedTestStruct(0, 0, Foo(0)));
-    sut.__for_each_field(&mut |offset, size| {
-        v.push((offset, size));
-    });
-
-    assert_that!(v, len 4);
-    assert_that!(v[0], eq(0, 8));
-    assert_that!(v[1], eq(8, 4));
-    assert_that!(v[2], eq(16, 8));
-    assert_that!(v[3], eq(24, 2));
-}
-
-#[test]
-pub fn field_offsets_and_sizes_are_correct_when_alignment_changes_inner_padding() {
-    #[repr(C)]
-    #[repr(align(16))]
-    #[derive(ZeroCopySend)]
-    struct AlignedU32(u32);
-
-    #[repr(C)]
-    #[derive(ZeroCopySend)]
-    struct SomeNamedStruct {
-        a: u8,
-        b: AlignedU32,
-    }
-
-    let mut v = Vec::new();
-    let sut = SomeNamedStruct {
-        a: 3,
-        b: AlignedU32(9),
-    };
-    sut.__for_each_field(&mut |offset, size| {
-        v.push((offset, size));
-    });
-
-    assert_that!(v, len 2);
-    assert_that!(v[0], eq(0, 1));
-    assert_that!(v[1], eq(16, 4));
-}
-
-#[test]
-pub fn field_offsets_and_sizes_are_correct_for_nested_structs() {
-    #[repr(C)]
-    #[derive(ZeroCopySend)]
-    struct SomeUnnamedStruct(u64);
-
-    #[repr(C)]
-    #[derive(ZeroCopySend)]
-    struct SomeNamedStruct {
-        a: u8,
-        b: i16,
-        c: SomeUnnamedStruct,
-    }
-
-    #[repr(C)]
-    #[derive(ZeroCopySend)]
-    struct NestedStruct {
-        a: SomeUnnamedStruct,
-        b: u32,
-        c: SomeNamedStruct,
-    }
-
-    let mut v = Vec::new();
-    let sut = NestedStruct {
-        a: SomeUnnamedStruct(32687),
-        b: 7,
-        c: SomeNamedStruct {
-            a: 1,
-            b: -4,
-            c: SomeUnnamedStruct(5),
-        },
-    };
-    sut.__for_each_field(&mut |offset, size| {
-        v.push((offset, size));
-    });
-
-    assert_that!(v, len 5);
-    assert_that!(v[0], eq(0, 8));
-    assert_that!(v[1], eq(8, 4));
-    assert_that!(v[2], eq(16, 1));
-    assert_that!(v[3], eq(18, 2));
-    assert_that!(v[4], eq(24, 8));
 }
